+++ sensors/sensor_sdc41.h
@@ -10,12 +10,13 @@ extern "C" {
 #endif
 
 #include <stdbool.h>
+#include <stdint.h>
 
 /* Basic measurement set for SDC41 (CO2 + T + RH) */
 typedef struct {
-    float co2_ppm;
-    float temperature_c;
-    float humidity_rh;
+    int32_t co2_ppm;
+    int32_t temperature_c;
+    int32_t humidity_rh;
 } sensor_sdc41_data_t;
 
 /* Initialize the sensor (I2C, pins, etc.). Return true on success. */
